refactor: web 코드들 역할별로 함수 및 파일 분리
