+++ src/utils/domUtils.js
@@ -1,0 +1,5 @@
+export function createDOMElement(tag, text = '') {
+  const element = document.createElement(tag);
+  element.textContent = text;
+  return element;
+}
